- added Cross-Site Request Forgery (CSRF) form protection - fixed redirect bug

+++ examples/forms/example1.php
@@ -74,7 +74,7 @@ $form->addSubmit('submit1', 'Send');
 $form['name']->addRule(Form::FILLED, 'Enter your name');
 
 $form['age']->addRule(Form::FILLED, 'Enter your age');
-$form['age']->addRule(Form::NUMERIC, 'Age must be numeric value');
+$form['age']->addRule(Form::INTEGER, 'Age must be numeric value');
 $form['age']->addRule(Form::RANGE, 'Age must be in range from %.2f to %.2f', array(9.9, 100));
 
 // conditional rule: if is email filled, ...
